Fix pixel order
The  specification  says "...samples  are  packed  into bytes  with  the
leftmost sample in  the high-order bits of a byte  followed by the other
samples for the  scanline.". This was not followed which  lead to images
using greyscale and  indexed colours with pixel width  smaller than full
byte to have the order of pixels reversed within the byte reversed.

Thanks to Samium Gromoff for the report and test case.